Update Solitare.c
Resolved previous commits' issue with copying and pasting everything and made it so that the computer works for you by reformating the print_card function to be implemented in a loop. The spacing looks iffy, but it's a good first start. I'm still toiling out the kinks on how I will keep track of what card has what value while the user manipulates the card order and moves stuff around. Currently thinking of implementing a 3D array (if possible) where the first two indices determine the card location on the screen, and the third index comprises a truth table determining if a card value can be shown to the user. Just a thought; I will have to think about other ways I could approach this before I move forward.