Just tidied up
I changed the card deck from an array to a matrix. The topmost row of the matrix signifies a card's value, and the second row designates a card suite. Then, the nested for loop in main( ) was revised to display card windows correctly; this was then moved to its own function called table_setup( ). Lastly, print_card( ) was modified to ensure that card suites were printed alongside a card's value with the proper color that correlates to the suite. I will now focus on toying around with mouse functions in ncurses to see how I'll move cards around the screen.